feat(ElevationControlNew): set sea level on both sides
also: refactor out drawBoundary function into its own module to guard
against a cyclical import of Redux store
