+++ client/src/boundary/index.ts
@@ -1,2 +1,3 @@
 export * from './boundary.js'
+export * from './draw.js'
 export * from './BoundaryComponent.js'
