+++ client/src/store/slices/street.ts
@@ -1,6 +1,7 @@
 import { createSlice } from '@reduxjs/toolkit'
 import { getSegmentInfo, getSegmentVariantInfo } from '@streetmix/parts'
 
+import { getBoundaryItem } from '~/src/boundary/boundary'
 import { getElevationValue } from '~/src/segments/elevation'
 import { getVariantString } from '~/src/segments/variant_utils'
 import { DEFAULT_SKYBOX } from '~/src/sky/constants'
@@ -428,6 +429,34 @@ const streetSlice = createSlice({
       },
     },
 
+    // This is like `setBoundaryElevation` but it changes only waterfront
+    // boundaries, and up to both at the same time
+    setSeaLevel(state, action: PayloadAction<number>) {
+      const value = action.payload
+
+      if (typeof value !== 'number' || Number.isNaN(value)) {
+        return
+      }
+
+      // function getBoundaryItem(str: string) {
+      //   return {
+      //     str,
+      //     waterfront: true
+      //   }
+      // }
+
+      // Set both boundaries if they are waterfronts
+      const left = getBoundaryItem(state.boundary.left.variant)
+      const right = getBoundaryItem(state.boundary.right.variant)
+
+      if (left.waterfront) {
+        state.boundary.left.elevation = value
+      }
+      if (right.waterfront) {
+        state.boundary.right.elevation = value
+      }
+    },
+
     // TODO: Buildings could be a child slice?
     addBuildingFloor(state, action: PayloadAction<BoundaryPosition>) {
       const position = action.payload
@@ -562,6 +591,7 @@ export const {
   addLocation,
   clearLocation,
   setBoundaryElevation,
+  setSeaLevel,
   addBuildingFloor,
   removeBuildingFloor,
   setBuildingFloorValue,
